Code optimisation and responsibility separation
