+++ InstagramStories/Source/Image Storer/IGCache.swift
@@ -1,0 +1,16 @@
+//
+//  ImageCache.swift
+//  InstagramStories
+//
+//  Created by Boominadha Prakash on 01/04/19.
+//  Copyright © 2019 DrawRect. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+let ONE_HUNDRED_MEGABYTES = 1024 * 1024 * 100
+
+class IGCache: NSCache<AnyObject, AnyObject> {
+    static let shared = IGCache()
+}
